refactor: replace NodeWebRtcEndpoint with composition
