+++ packages/network/test/integration/webrtc-multi-signaller.test.ts
@@ -5,17 +5,18 @@ import { Tracker, Event as TrackerEvent } from '../../src/logic/Tracker'
 import { PeerInfo } from '../../src/connection/PeerInfo'
 import { waitForEvent } from 'streamr-test-utils'
 import { Event as EndpointEvent } from '../../src/connection/IWebRtcEndpoint'
-import { NodeWebRtcEndpoint } from '../../src/connection/NodeWebRtcEndpoint'
 import { RtcSignaller } from '../../src/logic/RtcSignaller'
 import { NegotiatedProtocolVersions } from "../../src/connection/NegotiatedProtocolVersions"
+import { WebRtcEndpoint } from '../../src/connection/WebRtcEndpoint'
+import { NodeWebRtcConnectionFactory } from '../../src/connection/webRtcConnectionFactories'
 
 describe('WebRTC multisignaller test', () => {
     let tracker1: Tracker
     let tracker2: Tracker
     let trackerNode1: TrackerNode
     let trackerNode2: TrackerNode
-    let endpoint1: NodeWebRtcEndpoint
-    let endpoint2: NodeWebRtcEndpoint
+    let endpoint1: WebRtcEndpoint
+    let endpoint2: WebRtcEndpoint
 
     beforeEach(async () => {
         tracker1 = await startTracker({
@@ -46,10 +47,22 @@ describe('WebRTC multisignaller test', () => {
 
         const peerInfo1 = PeerInfo.newNode('node-1')
         const peerInfo2 = PeerInfo.newNode('node-2')
-        endpoint1 = new NodeWebRtcEndpoint(peerInfo1, ['stun:stun.l.google.com:19302'],
-            new RtcSignaller(peerInfo1, trackerNode1), new MetricsContext(''), new NegotiatedProtocolVersions(peerInfo1))
-        endpoint2 = new NodeWebRtcEndpoint(peerInfo2, ['stun:stun.l.google.com:19302'],
-            new RtcSignaller(peerInfo2, trackerNode2), new MetricsContext(''), new NegotiatedProtocolVersions(peerInfo2))
+        endpoint1 = new WebRtcEndpoint(
+            peerInfo1,
+            ['stun:stun.l.google.com:19302'],
+            new RtcSignaller(peerInfo1, trackerNode1),
+            new MetricsContext(''),
+            new NegotiatedProtocolVersions(peerInfo1),
+            NodeWebRtcConnectionFactory
+        )
+        endpoint2 = new WebRtcEndpoint(
+            peerInfo2,
+            ['stun:stun.l.google.com:19302'],
+            new RtcSignaller(peerInfo2, trackerNode2),
+            new MetricsContext(''),
+            new NegotiatedProtocolVersions(peerInfo2),
+            NodeWebRtcConnectionFactory
+        )
     })
 
     afterEach(async () => {
